+++ tests/java/main.ts
@@ -13,7 +13,11 @@ fs.rm(output, { recursive: true, force: true })
     await run({
       reader: defaultReader(input),
       writers: [markdownWriter(output), htmlWriter(output)],
-      plugins: [javaPlugin(output, { srcDir: src, ignoreAdditionalFiles: false })],
+      plugins: [javaPlugin(output, {
+        srcDir: src,
+        ignoreAdditionalFiles: false,
+        linkSrcDir: '../../src',
+      })],
     })
   })
   .then(async () => { await compareOutput(output, expected) })