+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added x-reference-type property to define the type of the reference
-- Added exclusion on application level
+- Added exclusions of validation errors on application level
+- Added support for links to java implementations using linkSrcDir property
 
 ### Changed